AT bool handeling
